feat: updated max total file size to 5MB
